fix bugs

+++ xxyp/src/main/java/com/xxyp/cache/CacheConstant.java
@@ -4,5 +4,6 @@ package com.xxyp.cache;
  * Created by jackeymm on 2018/3/29.
  */
 public class CacheConstant {
-    public static long cacheRefeshTime = 60*60*24*10L;
+//    public static long cacheRefeshTime = 60*60*24*10L;
+    public static int cacheRefeshTime = 60*60*24*10;
 }
